수정

+++ TeamProject_Malang-v2/src/main/java/com/malcom/malang/controller/NoticeController.java
@@ -6,6 +6,7 @@ import java.util.List;
 
 import javax.servlet.http.HttpSession;
 
+import com.malcom.malang.model.MemberVO;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -38,28 +39,50 @@ public class NoticeController {
 	}
 	
 	// --------------------- 게시글 삭제/업데이트 ----------------------------------
+
+	private boolean sessionCheck(Long bd_seq, HttpSession httpSession) {
+		MemberVO mVO = (MemberVO) httpSession.getAttribute("MEMBER");
+		BoardVO bVO = bService.findById(bd_seq);
+		int mb_role = mVO.getMb_role();
+		String mb_nickname = mVO.getMb_nickname();
+		String board_nickname = bVO.getBd_author();
+		if(mb_role == 2){
+			return false;
+		}
+		if(httpSession == null || mb_nickname != board_nickname){
+			return true;
+		}
+		return false;
+	}
 	
 	@RequestMapping(value="/view/notice/delete", method=RequestMethod.GET)
-	public String nDelete(Long bd_seq) {
-		
+	public String nDelete(Long bd_seq, HttpSession httpSession) {
+
+		if (sessionCheck(bd_seq, httpSession)) return "redirect:/";
+
 		bService.delete(bd_seq);
 		
 		return "redirect:/notice";
 	}
 	
 	@RequestMapping(value="/view/notice/update", method=RequestMethod.GET)
-	public String nUpdate(Long bd_seq, Model model) {
-		
+	public String nUpdate(Long bd_seq, Model model, HttpSession httpSession) {
+
+		if (sessionCheck(bd_seq, httpSession)) return "redirect:/";
+
 		BoardVO vo = bService.findById(bd_seq);
 		model.addAttribute("CTUP",vo);
 
 		
 		return "/board/write";
 	}
-	
+
 	@RequestMapping(value="/view/notice/update", method=RequestMethod.POST)
-	public String nUpdate(BoardVO vo, Model model) {
-	
+	public String nUpdate(BoardVO vo, Model model, HttpSession httpSession) {
+
+		Long bd_seq = vo.getBd_seq();
+		if (sessionCheck(bd_seq, httpSession)) return "redirect:/";
+
 		Integer result = bService.update(vo);
 		model.addAttribute("bd_seq", vo.getBd_seq());
 	
@@ -196,16 +219,5 @@ public class NoticeController {
 		
 		return "/board/notice";
 	}
-	
-	
-//	private List<BoardVO> subDate(List<BoardVO> bList) {
-//	
-//		for(int i = 0 ; i < bList.size() ; i++) {
-//			String date = bList.get(i).getBd_date().substring(0, 10);
-//			bList.get(i).setBd_date(date);
-//		}
-//		
-//		return bList;
-//	}
 
 }
